lo que hace falta para que compile en mi casa

+++ srcs/utils/ft_open_file.c
@@ -6,7 +6,7 @@
 /*   By: vicgarci <vicgarci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 14:07:50 by vicgarci          #+#    #+#             */
-/*   Updated: 2023/05/23 16:17:27 by vicgarci         ###   ########.fr       */
+/*   Updated: 2023/07/19 18:02:01 by vicgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,11 +48,11 @@ aleatoria
 */
 char	*ft_open_file(t_shell *shell)
 {
-	int			seed;
+	uintptr_t	seed;
 	char		*s;
 
 	s = NULL;
-	seed = ((((int)&seed) / 4) >> 2) * -1;
+	seed = ((((uintptr_t)&seed) / 4) >> 2) * -1;
 	while (true)
 	{
 		s = ft_itoa(seed);
